working encoding of documents: step 1 in having a mongo driver

+++ typeinfo.h
@@ -18,7 +18,7 @@ namespace bson
   
   char to_char(const TypeInfo & ti)
   {
-    return static_cast<char>(ti);
+    return static_cast<char>((int)ti);
   }
   
   //forward declaration
